Use blueprints not controllers, removed rogue pylons refs
