--- ckanext/query_dois/controllers/__init__.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-# encoding: utf-8
-#
-# This file is part of ckanext-query-dois
-# Created by the Natural History Museum in London, UK
-
-try:
-    import pkg_resources
-
-    pkg_resources.declare_namespace(__name__)
-except ImportError:
-    import pkgutil
-
-    __path__ = pkgutil.extend_path(__path__, __name__)
