+++ ckanext/query_dois/routes/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# encoding: utf-8
+#
+# This file is part of ckanext-query-dois
+# Created by the Natural History Museum in London, UK
+
+from . import query_doi
+
+blueprints = [query_doi.blueprint]
